Fixed issue with map names
